requirements.txt
numpy~=1.19.2
pyyaml
matplotlib~=3.2.2
pandas~=1.3.4
pytorch-lightning>=1.5
scikit-learn~=0.24.1
neptune-client
umap-learn

requests~=2.25.1
torchvision~=0.10.1
sklearn~=0.0
torchmetrics~=0.4.1
pillow~=8.1.2
scipy~=1.6.2
anndata~=0.7.6
colorcet~=2.0.6
scanpy~=1.8.1
leidenalg~=0.8.3
seaborn~=0.11.1
pytest~=6.2.3
setuptools~=49.6.0
